refactor thl-front, fix 'forms' render problem, change cypress test to match changes
